Improve handling of bugs listed against src:foo
Rather than mapping "src:foo" to "foo" whilst building the bug hashes,
we store the src: bugs "as is" in the hash, and then include them in the
list of relevant bugs when building the list for a source package.

This avoids a situation where a bug filed against "src:foo" can impede
the migration of the binary package "foo", built from a different source
package.

Signed-off-by: Adam D. Barratt <adam@adam-barratt.org.uk>