Cleanup old binaries even when they are still in unstable
Old binaries in testing are usually old versions of libraries that where
smooth-updated. These should only be kept in testing as long as they are
needed to install other packages in testing. When they are no longer needed,
they can be removed, even if the old library is still in unstable (where it
might be needed for reverse dependencies which are not in testing).

Signed-off-by: Ivo De Decker <ivodd@debian.org>