Remove redundant notes about arch:all packages
In an architecture only migration, we currently see two notes for
every arch:all packages:

 * Ignoring "new" arch:all package
 * Ignoring "removal" of arch:all package

But a closer look at the situation is that the arch:all packages is
generally the same in both testing and the source suite.

This commit removes these notes when the arch:all is the same in both
suites.

Signed-off-by: Niels Thykier <niels@thykier.net>