Remove dead case in undo code
The doop_source method does not generate any /binary/ undo items,
where the package name starts with "-".  This fact allows us to remove
a branch in the undo code.  This also happens to be fortunate because
the call to "inst_tester.remove_testing_binary" in that branch happens
to be wrong.

Signed-off-by: Niels Thykier <niels@thykier.net>