britney: Use direct and indirect affected packages for nuninst checks
Previously, we split affected into a "reverse relations with a
dependency on the migrated/change package" set and a set of "the
rest".  With this commit we change the sets into a set of packages
with a direct relation on the migrated/changed package and the full
set of all affected.

The change has no effect on Britney's results as the first set is only
used as an optimization to reject some items faster.

Signed-off-by: Niels Thykier <niels@thykier.net>