Add invalidated-by-other-package field in excuses.yaml
The other fields related to invalidation are currently not deterministic and
depend on the order in which the excuses are invalidated.

The new invalidated-by-other-package field is just a boolean which shows the
excuse was invalidated by some other package for some reason (without
specifying the reason). It does not depend on the order of processing and
should be deterministic.

Signed-off-by: Ivo De Decker <ivodd@debian.org>