autopkgtest: Extract some ("boring") loops into a function
When we convert legacy results in the autopkgtest-results.cache file,
we are only touching leaf results.  By moving the loops into a
function, we can remove 2-3 levels of ("redundant") nesting.  This in
turn makes it more clear what is relevant in the conversion.

This same also holds in save_state when we convert an Enum to a
string.

Signed-off-by: Niels Thykier <niels@thykier.net>